02_StreamLit/03_ChatBot/requirements.txt
streamlit
streamlit_authenticator
groq
sqlmodel
python-dotenv